Merged blowfish.pas into bcrypt.pas
- Removed dependancy on Blowfish.pas. Copied relavent ECB function,
state variable, and digits of PI
- We now recognize version "2" and well as anything version
"2"+[letter].
- We now use OS CryptGenRandom for salt generation.
If it fails, we fall back to a type 4 ("random") UUID.
Salt doesn't have to be cryptographically strong, just different.
- Added: CheckPassword functions now take "PasswordRehashNeeded" out
parameter. The value will contain True if the password needs to be
rehashed.
- Fixed up compiler defines so that we work on Delphi 5, Delphi 7, and
Delphi XE5 (the only versions of Delphi i use)
- Added the fast SelfTests into their own DUnit test; the ones that are
too slow are still kept in the public section
- Added granular abilty to disable BCrypt unit tests