Apply unicode NFKC as per NIST SP 800-63B
In accordance with the recommendations of NIST SP 800-63B, we now apply
KC normalization to the password.
Choice was between NFKC and NFKD; i think compatibility *composition*
(rather than decomposition) is better.
I don't know if anyone else has any standard yet on which way to go. I
made a test to match it.